requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
mlflow==2.7.1
datasets==2.14.5
huggingface-hub==0.17.3
streamlit==1.28.1
joblib==1.3.2
python-dotenv==1.0.0
pyyaml==6.0.1
